[강운덕] [PROFILER-31] rpc call 호출 api 추가로 인해 testcase 실패하여 commit함.
git-svn-id: http://svn.bds.nhncorp.com/pe/hippo-tomcat-profiler/trunk@3981 84d0f5b1-2673-498c-a247-62c4ff18d310

+++ src/main/java/com/nhn/pinpoint/profiler/monitor/metric/LongAdderHistogram.java
@@ -31,9 +31,9 @@ public class LongAdderHistogram implements Histogram {
         return serviceType;
     }
 
-    public void addResponseTime(int millis) {
+    public void addResponseTime(long millis) {
         HistogramSchema schema = serviceType.getHistogramSchema();
-        final HistogramSlot histogramSlot = schema.findHistogramSlot(millis);
+        final HistogramSlot histogramSlot = schema.findHistogramSlot((int) millis);
         final SlotType slotType = histogramSlot.getSlotType();
         switch (slotType) {
             case FAST:
